+++ libpromises/cf3globals.c
@@ -49,7 +49,7 @@ Item *PROCESSTABLE = NULL; /* GLOBAL_X */
 /* Internal data structures                                                  */
 /*****************************************************************************/
 
-int LASTSEENEXPIREAFTER = SECONDS_PER_WEEK; /* GLOBAL_P */
+long LASTSEENEXPIREAFTER = SECONDS_PER_WEEK; /* GLOBAL_P */
 
 char POLICY_SERVER[CF_MAX_IP_LEN] = ""; /* GLOBAL_X */
